Store the command allocator in the private data of the command list so that it can be retrieved from the command list when it is executed on the command queue.
Storing the command allocator in the private data of the command list allows the command allocator to be placed on an allocator queue when the command list is submitted to the command queue. The allocator is not reused again until the fence value has been reached.
Doing this also made the CommandList class obsolete and is removed in this commit.
Also added a function to report any live objects at application exit.